Model_deployment/requirements.txt
streamlit
numpy
pandas
fastapi[all]
python-dotenv